update markers module to stick to markers
